Reorganize subpackage
